+++ AoDDownloader/utils.py
@@ -6,7 +6,7 @@ import AoDDownloader as AoD
 from .config import Config
 
 
-def create_downloader() -> AoD.AoDDownloader:
+def create_downloader(password) -> AoD.AoDDownloader:
     config = Config()
 
     if config.username:
@@ -17,14 +17,18 @@ def create_downloader() -> AoD.AoDDownloader:
                 click.echo(
                     "Keyring is not accessible.\n \"pip install dbus-python\" might fix this problem.")
         else:
-            config.password = click.prompt('Password', hide_input=True)
+            config.password = password if password else click.prompt('Password', hide_input=True)
     return AoD.AoDDownloader(config=config)
 
 
-def create_login(use_keyring: bool = True):
+def create_login(use_keyring: bool = True, username: str = '', password: str = ''):
     config = Config()
-    config.username = click.prompt('Username')
-    config.password = click.prompt('Password', hide_input=True)
+    if(not username or not password):
+        config.username = click.prompt('Username')
+        config.password = click.prompt('Password', hide_input=True)
+    else:
+        config.username = username
+        config.password = password
     config.keyring = use_keyring
     try:
         aod = AoD.AoDDownloader(config=config)
@@ -32,7 +36,7 @@ def create_login(use_keyring: bool = True):
         click.echo(e)
         return
     if not aod.signed_in:
-        click.echo("Login fehlgeschlagen.")
+        click.echo("Login failed.")
         return
 
     if use_keyring:
@@ -44,6 +48,7 @@ def create_login(use_keyring: bool = True):
 "pip install dbus-python" might fix this problem.
 Or use --no-keyring to enter password every time and disable keyring""")
     config.write()
+    click.echo("Login successfull.")
 
 
 def remove_login():