+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 - add integer only support
 - add escape of ffmpeg strings
 - add progressbar when parsing episodes
+- add direct arguments for login and download
+- add `no-buffer-output` option to get a not buffered output
+- try signing in when 403 is thrown
 
 # 1.0.0rc5
 - Use no temp file on windows because of permission errors